Fix dynamo utils split review regressions
